// src/classes/models/session.ts
export interface IRow {
    bestnlapsnum: number;
    bestnlapstime: number;
    helm_color1: string;
    bestlapnum: number;
    car_pattern: number;
    suit_pattern: number;
    bestquallaptime: number;
    newcpi: number;
    simsesname: string;
    hostid: string;
    ccName: string;
    license_change_road: number;
    oldirating: number;
    division: number;
    max_pct_fuel_fill: number;
    reasonout: string;
    oldttrating: number;
    carnumberfont: number;
    sessionstarttime: any;
    ccNameShort: string;
    newirating: number;
    wheel_chrome: number;
    helm_color3: string;
    wheel_color: string;
    oldcpi: number;
    helm_color2: string;
    weight_penalty_kg: number;
    multiplier: number;
    groupid: number;
    clubpoints: number;
    vehiclekeyid: number;
    dropracepoints: number;
    officialsession: number;
    newsublevel: number;
    car_color1: string;
    licensecategory: string;
    track_catid: number;
    car_color3: string;
    bestquallapat: any;
    car_color2: string;
    suit_color2: string;
    clubname: string;
    suit_color1: string;
    aggchamppoints: number;
    suit_color3: string;
    custid: number;
    interval: number;
    oldlicenselevel: number;
    avglap: number;
    classinterval: number;
    simsesnum: number;
    quallaptime: number;
    track_category: string;
    carclassid: number;
    newttrating: number;
    restrictresults: string;
    clubid: number;
    carnum: string;
    startpos: number;
    carid: number;
    lapscomplete: number;
    car_number_color1: string;
    car_number_color2: string;
    licensegroup: number;
    car_number_color3: string;
    pos: number;
    bestlaptime: number;
    incidents: number;
    champpoints: number;
    heatinfoid: number;
    simsestypename: string;
    finishpos: number;
    license_change_oval: number;
    carsponsor2: number;
    helm_pattern: number;
    carsponsor1: number;
    divisionname: string;
    carnumberslant: number;
    reasonoutid: number;
    finishposinclass: number;
    optlapscomplete: number;
    lapslead: number;
    oldsublevel: number;
    displayname: string;
    league_points: string;
    evttypename: string;
    clubshortname: string;
    bestquallapnum: number;
    newlicenselevel: number;
}

export interface ISession {
    nleadchanges: number;
    weather_wind_dir: number;
    league_season_id: string;
    privatesessionid: number;
    race_week_num: number;
    sessionid: number;
    pointstype: string;
    season_shortname: string;
    weather_fog_density: number;
    simsestype: number;
    specialeventtype: number;
    ncautions: number;
    eventlapscomplete: number;
    cornersperlap: number;
    simulatedstarttime: string;
    seasonid: number;
    weather_type: number;
    rubberlevel_practice: number;
    season_quarter: number;
    driver_change_rule: number;
    evttype: number;
    series_name: string;
    subsessionid: number;
    leavemarbles: number;
    rubberlevel_qualify: number;
    weather_rh: number;
    weather_wind_speed_value: number;
    weather_wind_speed_units: number;
    weather_temp_units: number;
    weather_temp_value: number;
    maxweeks: number;
    min_team_drivers: number;
    driver_change_param1: number;
    driver_change_param2: number;
    cautiontype: number;
    sessionname: string;
    nlapsforqualavg: number;
    season_name: string;
    seriesid: number;
    start_time: string;
    timeofday: number;
    leagueid: string;
    trackid: number;
    specialeventtypetext: string;
    rubberlevel_race: number;
    eventstrengthoffield: number;
    weather_skies: number;
    catid: number;
    rserv_status: string;
    season_year: number;
    track_name: string;
    series_shortname: string;
    max_team_drivers: number;
    rubberlevel_warmup: number;
    nlapsforsoloavg: number;
    ncautionlaps: number;
    track_config_name: string;
    driver_changes: number;
    eventavglap: number;
    rows: IRow[];
}

export interface IRaceSession {
    customer_id: string;
    id: number;
    session: ISession;
}